Disable CDN setup
We have broken the PaaS. Let's be nice and lay off them for a short
while, whilst they investigate the problem and perhaps fix it.

The problem is, that without the domain actually pointing at the CDN, it
will never be capable of resolving the record and assigning SSL
certificate. This essentially means, it will attempt to do that
indefinitely, causing PaaS to self-DDoS...

Sorry PaaSians!