REP-355: Scale apps post-deploy
Since we're deploying so many apps on a PaaS plan of 10GB, we need to
limit the amount of memory allocated to each app.

The app manifest currently gets generated by the registers-cli which
makes it a bit messy to alter any resource allocation prior to
deploying. Instead this adds a task which we can use to make any changes
to the app right after deploying.