REP-355: Shorten PaaS resource names
PaaS has a limit of 50 characters for some resource names, so to keep
things simple we'll just call things `app`, `route-service` and
`route-service-binding` for the time being.

I've stopped using the cf-resource since it doesn't give us the
flexibility we need.

We can put the cdn stuff back once we're ready to switch DNS.