Build and deploy single app
This defines a single pipeline for publishing all the registers as a
single PaaS app using the nginx buildpack.

A modified nginx config lives in this repo to support serving multiple
registers from the same PaaS app.

Compiling the registers is a relatively slow process, so the
static files resulting from the `build` job are output as an archive in
S3 to be pulled and deployed to PaaS by the `app` job. Here is a rundown
of the jobs in this pipeline:

- `update`: updates the pipeline
- `build`: compiles `.rsf` from the registry-data git repo into static
CSV and JSON files
- `app`: pushes the app resources to PaaS
- `route-service`: tests the Go route service and pushes it to PaaS
- `routes`: binds the registers routes to the route service
- `test`: runs the conformance tests on the registers